(PC-3300): Extract formatAndOrderVenues in specific util file
